Code refactor & profile page final touches

+++ backend/services/serializers.py
@@ -3,7 +3,9 @@ from rest_framework import serializers
 class WebviewSerializer(serializers.Serializer):
     type = serializers.CharField()
     address = serializers.CharField()
-    radius = serializers.IntegerField()
+
+    # Radius should be float (not int)
+    radius = serializers.FloatField()
